Improvements to "Change dir" option
Settings are now always stored in the script directory, and changing dir will restart the game and load the new set of settings. It's basically script profiles now.
autorunsettings.txt in the base dir now only stores "DIR". Previous settings that may have been there are migrated to the new location.

For the 99.9% of users that don't use this feature, nothing changes, settings are still stored in scripts/autorunsettings.lua