Add modulepack.lua
modulepack.lua packs multi-module scripts into a single file that can be publishes on the Script Manager. modulepack.lua expects the path to a modulepack.conf file as its first argument, which lists a module on each line formatted appropriately for a call to require, to be found and made available for loading by other modules. The first line in the file is treated as the main module, and the resulting script calls the "run" function of this main module to run the multi-module script.

modulepack.lua outputs the resulting script to stdout by default, but if as the second argument modulepack.lua receives the string "run", it runs the resulting script instead.

All versions of Lua supported by TPT are also supported by this script. This will at some point replace the mess that is TPTMP's module packer script.